+++ src/main/java/nextstep/auth/AuthArgumentResolver.java
@@ -1,5 +1,6 @@
 package nextstep.auth;
 
+import nextstep.auth.annotation.Authenticated;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
