+++ src/main/java/nextstep/auth/annotation/Authenticated.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package nextstep.auth.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
