+++ src/main/java/nextstep/reservation/ReservationController.java
@@ -1,6 +1,6 @@
 package nextstep.reservation;
 
-import nextstep.auth.Authenticated;
+import nextstep.auth.annotation.Authenticated;
 import nextstep.auth.LoginUser;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
